Add test xtask to run all tests and generate/review snapshots
It looks like cargo-insta properly runs all targets and then the
doctests out of the box, which is nice to have everything all together
in one program. This will also generate any applicable snapshots and
then prompt you to review them if there are changes. Nice shortcut!
