+++ xtask/src/dev.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use xshell::Shell;
+use xshell::{cmd, Shell};
 
 use crate::commands::cargo_cmd;
 use crate::utils::{project_root, verbose_cd};
@@ -16,6 +16,47 @@ pub fn cargo_watch(config: &Config) -> Result<()> {
             "--why",
             "-x",
             "clippy --locked --all-targets --all-features",
+        ];
+        cmd.args(args).run()?;
+    }
+
+    Ok(())
+}
+
+pub fn install_rust_deps(config: &Config) -> Result<()> {
+    let sh = Shell::new()?;
+    verbose_cd(&sh, project_root());
+
+    cmd!(sh, "rustup component add clippy rustfmt").run()?;
+
+    let cmd_option = cargo_cmd(config, &sh);
+    if let Some(cmd) = cmd_option {
+        let args = vec![
+            "install",
+            "cargo-insta",
+            "cargo-llvm-cov",
+            "cargo-nextest",
+            "cargo-watch",
+        ];
+        cmd.args(args).run()?;
+    }
+
+    Ok(())
+}
+
+pub fn test_with_snapshots(config: &Config) -> Result<()> {
+    let sh = Shell::new()?;
+    verbose_cd(&sh, project_root());
+
+    let cmd_option = cargo_cmd(config, &sh);
+    if let Some(cmd) = cmd_option {
+        let args = vec![
+            "insta",
+            "test",
+            "--all-features",
+            "--test-runner",
+            "nextest",
+            "--review",
         ];
         cmd.args(args).run()?;
     }
